# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP V2 API",
        "description": "Comprehensive ERP System API Documentation with Multi-tenancy and Role-based Access Control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Get all users",
                "operationId": "9ded17315e13756cd593b5840cd39a26",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Create a new user",
                "operationId": "f6f90d5fbbe24f66eb2b43901e0d3f38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Get user by ID",
                "operationId": "c6ac2250cb22823f621026268b66f471",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Update user",
                "operationId": "22692373e5a35c609175c1e6441b58ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Delete user",
                "operationId": "fdbad6312957544572f3b379b96abad0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/subscriptions/my-subscription": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get user's active subscription",
                "operationId": "df4e477c659c5ab361db45719030f276",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subscriptions/subscribe": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Subscribe to a plan",
                "operationId": "cee38457a3adcffb4b35819ef00a7989",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubscribeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subscriptions/subscriptions/{id}/unsubscribe": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Cancel subscription",
                "operationId": "70338943fa35f95b712c42928b042659",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnsubscribeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription cancelled"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subscriptions/subscriptions/{id}/renew": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Renew subscription",
                "operationId": "f5b2e6100dddb36231982c41b37bbdaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription renewed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subscriptions/my-subscription-history": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get subscription history",
                "operationId": "8dd03c67c669efa19197581c439c06e6",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubscribeRequest": {
                "required": [
                    "plan_id"
                ],
                "properties": {
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_trial": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "UnsubscribeRequest": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "example": "Too expensive",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreUserRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "branch_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "avatar_path": {
                        "type": "string",
                        "example": "/avatars/user.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "admin"
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "newpassword123",
                        "nullable": true
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "newpassword123",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "branch_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "avatar_path": {
                        "type": "string",
                        "example": "/avatars/user.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "manager"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}